python-weather/requirements.txt
falcon
gunicorn
requests
python-dateutil
-e libs/docker_secrets
